style(ListBox): UXD-1692 render checkbox
