+++ packages/ListBox/package.json
@@ -8,6 +8,7 @@
     "@babel/runtime-corejs2": "^7.3.1",
     "@paprika/button": "^1.1.10",
     "@paprika/constants": "^1.0.0",
+    "@paprika/checkbox": "^1.0.20",
     "@paprika/helpers": "^2.2.0",
     "@paprika/icon": "^2.1.5",
     "@paprika/input": "^4.0.16",
