+++ packages/ListBox/src/index.d.ts
@@ -150,6 +150,8 @@ declare namespace ListBox {
     id: string;
     /** Internal prop, which shouldn't be documented */
     preventDefaultOnSelect?: boolean;
+    /** Let the user to select multiple options at same time */
+    isMulti?: boolean;
   }
 }
 declare namespace ListBox {
